feat: Navigate into taps with persist states for each views
