+++ cinemapedia/lib/config/router/app_router.dart
@@ -2,21 +2,29 @@ import 'package:go_router/go_router.dart';
 
 import 'package:cinemapedia/presentation/screens/screens.dart';
 
-final appRouter = GoRouter(initialLocation: '/', routes: [
+final appRouter = GoRouter(initialLocation: '/home/0', routes: [
   GoRoute(
-      path: '/',
+      path: '/home/:page',
       name: HomeScreen.name,
-      builder: (context, state) => const HomeScreen(),
+      builder: (context, state) {
+        final pageIndex = int.parse(state.pathParameters['page'] ?? '0');
+
+        return HomeScreen(pageIndex: pageIndex);
+      },
       routes: [
         GoRoute(
-           path: 'movie/:id',
-            name: MovieScreen.name,
-            builder: (context, state) {
-              //How to get the parameter:
-              final movieId = state.pathParameters['id'] ?? 'no-id';
-              return MovieScreen(
-                movieId: movieId,
-              );
-            }),
+          path: 'movie/:id',
+          name: MovieScreen.name,
+          builder: (context, state) {
+            final movieId = state.pathParameters['id'] ?? 'no-id';
+
+            return MovieScreen(movieId: movieId);
+          },
+        ),
       ]),
+      // Apply a redirect
+  GoRoute(
+    path: '/',
+    redirect: (_, __) => '/home/0',
+  ),
 ]);
