+++ cinemapedia/lib/presentation/screens/movies/home_screen.dart
@@ -1,96 +1,30 @@
-import 'package:cinemapedia/presentation/providers/providers.dart';
 import 'package:flutter/material.dart';
+
 import 'package:cinemapedia/presentation/widgets/widgets.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:cinemapedia/presentation/views/views.dart';
 
 class HomeScreen extends StatelessWidget {
   static const name = 'home-screen';
+  final int pageIndex;
 
-  const HomeScreen({super.key});
+  const HomeScreen({super.key, required this.pageIndex});
 
-  @override
-  Widget build(BuildContext context) {
-    return const Scaffold(
-      body: _HomeView(),
-      bottomNavigationBar: CustomBottomNavigation(),
-    );
-  }
-}
-
-class _HomeView extends ConsumerStatefulWidget {
-  const _HomeView();
-
-  @override
-  _HomeViewState createState() => _HomeViewState();
-}
-
-class _HomeViewState extends ConsumerState<_HomeView> {
-  @override
-  void initState() {
-    super.initState();
-
-    ref.read(nowPlayingMoviesProvider.notifier).loadNextPage();
-    ref.read(popularMoviesProvider.notifier).loadNextPage();
-    ref.read(topRatedMoviesProvider.notifier).loadNextPage();
-    ref.read(upcomingMoviesProvider.notifier).loadNextPage();
-  }
+// List of Widgets for the taps into TabBottomNavigator
+  final viewRoutes = const <Widget>[
+    HomeView(),
+    SizedBox(), // <--- categorias View
+    FavoritesView(),
+  ];
 
   @override
   Widget build(BuildContext context) {
-    final initialLoading = ref.watch(initialLoadingProvider);
-    if (initialLoading) return const FullScreenLoader();
-    // Observers for differents cases
-    final slideShowMovies = ref.watch(moviesSlideshowProvider);
-    final nowPlayingMovies = ref.watch(nowPlayingMoviesProvider);
-    final popularMovies = ref.watch(popularMoviesProvider);
-    final topRatedMovies = ref.watch(topRatedMoviesProvider);
-    final upcomingMovies = ref.watch(upcomingMoviesProvider);
-
-    return CustomScrollView(slivers: [
-      // Sliver with appbar for the scrollview.
-      const SliverAppBar(
-        floating: true,
-        flexibleSpace: FlexibleSpaceBar(
-          // add the customn app bar for manage the header into the scroll view.
-          title: CustomAppbar(),
-        ),
+    return Scaffold(
+      // The widget IndexedStact persist the state for the current widget.
+      body: IndexedStack(
+        index: pageIndex,
+        children: viewRoutes,
       ),
-
-      // The slivers are a widget with beheivors for the scrollviews
-      SliverList(
-          delegate: SliverChildBuilderDelegate((context, index) {
-        return Column(
-          children: [
-            MoviesSlideshow(movies: slideShowMovies),
-            MovieHorizontalListview(
-                movies: nowPlayingMovies,
-                title: 'En cines',
-                subTitle: 'Lunes 13',
-                //  read the privider via ref.read ...
-                loadNextPage: () =>
-                    ref.read(nowPlayingMoviesProvider.notifier).loadNextPage()),
-            MovieHorizontalListview(
-                movies: upcomingMovies,
-                title: 'Próximamente',
-                subTitle: 'En este mes',
-                loadNextPage: () =>
-                    ref.read(upcomingMoviesProvider.notifier).loadNextPage()),
-            MovieHorizontalListview(
-                movies: popularMovies,
-                title: 'Populares',
-                // subTitle: '',
-                loadNextPage: () =>
-                    ref.read(popularMoviesProvider.notifier).loadNextPage()),
-            MovieHorizontalListview(
-                movies: topRatedMovies,
-                title: 'Mejor calificadas',
-                subTitle: 'Desde siempre',
-                loadNextPage: () =>
-                    ref.read(topRatedMoviesProvider.notifier).loadNextPage()),
-            const SizedBox(height: 10),
-          ],
-        );
-      }, childCount: 1)),
-    ]);
+      bottomNavigationBar: CustomBottomNavigation(currentIndex: pageIndex),
+    );
   }
 }
