+++ cinemapedia/lib/presentation/widgets/shared/custom_bottom_navigation.dart
@@ -1,27 +1,40 @@
 import 'package:flutter/material.dart';
-
+import 'package:go_router/go_router.dart';
 
 class CustomBottomNavigation extends StatelessWidget {
-  const CustomBottomNavigation({super.key});
+  final int currentIndex;
+  const CustomBottomNavigation({super.key, required this.currentIndex});
+
+// When de user tap on the button is for index so define what route need for each tap
+  void onItemTapped(BuildContext context, int index) {
+    switch (index) {
+      // Main Screen
+      case 0:
+        context.go('/home/0');
+        break;
+      //Categories Screen
+      case 1:
+        context.go('/home/1');
+        break;
+      //Favorites Screen
+      case 2:
+        context.go('/home/2');
+        break;
+    }
+  }
 
   @override
   Widget build(BuildContext context) {
     return BottomNavigationBar(
-      elevation: 0,
-      items: const [
-        BottomNavigationBarItem(
-          icon: Icon( Icons.home_max ),
-          label: 'Inicio'
-        ),
-        BottomNavigationBarItem(
-          icon: Icon( Icons.label_outline ),
-          label: 'Categorías'
-        ),
-        BottomNavigationBarItem(
-          icon: Icon( Icons.favorite_outline ),
-          label: 'Favoritos'
-        ),
-      ]
-    );
+        elevation: 0,
+        currentIndex: currentIndex,
+        onTap: (value) => onItemTapped(context, value),
+        items: const [
+          BottomNavigationBarItem(icon: Icon(Icons.home_max), label: 'Inicio'),
+          BottomNavigationBarItem(
+              icon: Icon(Icons.label_outline), label: 'Categorías'),
+          BottomNavigationBarItem(
+              icon: Icon(Icons.favorite_outline), label: 'Favoritos'),
+        ]);
   }
 }
